Moved base test class to common.py
