+++ minorminer/layout/tests/test_find_embedding.py
@@ -7,7 +7,7 @@ import numpy as np
 
 import minorminer.layout as mml
 
-from . import TestLayoutPlacement
+from .common import TestLayoutPlacement
 
 
 class TestFindEmb(TestLayoutPlacement):
